Fixing windows compilation

+++ src/api/mem_streambuf.cpp
@@ -90,7 +90,7 @@ streamsize mem_streambuf::xsgetn( char_type *buf, std::streamsize size ) {
   if ( size > remaining )
     size = remaining;
   ::memcpy( buf, gptr(), static_cast<size_t>( size ) );
-  gbump( size );
+  gbump( static_cast<int>(size) );
   return size;
 }
 
@@ -99,7 +99,7 @@ streamsize mem_streambuf::xsputn( char_type const *buf, streamsize size ) {
   if ( size > remaining )
     size = remaining;
   ::memcpy( pptr(), buf, static_cast<size_t>( size ) );
-  pbump( size );
+  pbump( static_cast<int>(size) );
   return size;
 }
